Performance improvement
Secbox checks if you are connected to the SUSE internal network by
probing a well-known ssh server. In case you are not, commands might be
delayed due to a slow connection timeout. This commit force the check to
fail in case no SYN+ACK or RST packet is received after one second from
the first SYN.