Add options to manually enable/disable tty
The php interpreter included in secbox-image [0] is mainly used
in an interactive way, but there might be circumstances where you want
to run it in a non-interactive way [1]. Hence, I added two new flags:
`--no-tty` and `--tty` to force the `exec` behavior. Most of the time
none of them is required because secbox automatically detects when an
interactive terminal is required and when it's not.

[0] https://gitlab.suse.de/security/secbox-image
[1]
https://github.com/StayPirate/dotfiles/commit/58595aee043b16ae57cfb8ec067c81a0fc00c318