Dynamically load environment variables
Fix: #8

Before this commit, all the environment variables were initialized at
container creation. Because of that, it was not possible to have
different values in the subsequent execution, and this misbehaved in
some circumstances, for instance in openSUSE Tumbleweed the
$SSH_AUTH_SOCK value changes across each boot and the container is not
able to start because it tries to bind mount a file that no longer
exists. The design I implemented here is very similar to what secbox
already does with resolv.conf [0]. Secbox now manages its own env-file
which is updated and loaded every time an exec is called.

[0] https://git.io/JR9fo