Replace conainer w/one based on latest image built
It might happen that you want to replace the current container with a
new one based on the latest container-image built. For instance, this
can be handy when a new version of a package (installed in the image) is
released but the image itself hasn't got any version bump.
You can now use `secbox --update-container --force` to pull the latest
image built and replace the container with one based on that image.

Fixes #14