Fix a bug during the update process
The bug was reported by @gsoonu and was triggered when the container
was update with the option -f "--update-container -f" the old image was
not deleted and the container was recreated from it. This could be only
experienced when the newer image tag was numerically lower than the
previous one, like: 2.10 and 2.9.