fix(sshfs): Replace wotan with dist.suse.de after its decommission
Wotan was used in secbox to automatically provide network shares
required by some of our internal tools. It's been decomissioned by the
infra team, so I'm replacing it with dist.suse.de which in turn provides
the same network shares.

This is a breaking change because it requires users to configure again
their ~/.ssh/config files. See the README.md for instructions.