fix(secbox): Pull most recent upstream tag when no local tag found
A colleague reported an issue triggered when no container and no local
image exist. Secbox considers an empty local tag, and it tries to use
it to create the container:

  podman container create registry.suse.de/non_public/maintenance/security/container/containers/secbox:

TBH I was not able to reproduce it. But, by adding the remote most recent
tag (when the local tag is empty), we can force podman to pull the
missing image at container creation time.